orchestrator/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
websockets==15.0.1
httpx==0.28.1
python-multipart==0.0.20
faster-whisper==1.1.1
piper-tts==1.2.0
